Added Models and Set Controls For Different Cameras
We added models we want to use in the future.

We also set controls for orthographic to move on the x and y axes with
the arrow keys, and moving in the x and z axes in the perspective camera
with the arrow keys.
The shift modifier is still an option to move in the third axis.