Added Code and Comments for the RigidBody.h
It is a header with all of the calculations needed for the object to work.

Next step is to integrate it into our existing code, which can be done by editing our AppClass, and Bounding Object Manager Classes. We also need to see if we can edit the Mesh manager slightly with Alberto's Permission to include a rigid body member to hold all of the data and do calculations. If we cannot, I can make it a singleton and edit the values read in from the object as opposed to each object having one. This would require that we make a manager though, so I will look into that just in case.